+++ src/main/kotlin/calculator/ConsoleLauncer.kt
@@ -1,5 +1,7 @@
 package calculator
 
+import calculator.app.StringCalculator
+
 fun main() {
     val calculator = StringCalculator()
     calculator.calculate(readlnOrNull())
