+++ src/test/kotlin/calculator/model/OperatorTest.kt
@@ -1,3 +1,10 @@
 package calculator.model
 
-class OperatorTest
+import org.junit.jupiter.api.Test
+
+class OperatorTest {
+
+    @Test
+    fun `사칙연산 기호가 아닌 경우 IllegalArgumentException throw`() {
+    }
+}
